backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
PyMuPDF==1.24.4
pytest==8.2.2
httpx==0.27.0
typing_extensions>=4.12.0
opentelemetry-api>=1.25.0
opentelemetry-sdk>=1.25.0
opentelemetry-exporter-otlp-proto-http>=1.25.0
langsmith>=0.1.56
